requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
filelock==3.0.12
gast==0.3.3
google==3.0.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
msgpack==1.0.0
multidict==4.7.6
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
protobuf==3.12.2
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
PyYAML==5.3.1
ray==0.8.6
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
